Progress on shrinking counter examples
When a property fails, we want to shrink the counter example to
something succinct which still fails the test.  shrink/2 handles the
basic shrinking step.  I've refined it substantially based on Haskell's
instances for Test.QuickCheck.Arbitrary and by playing around in the
Prolog top level.

The next step is to recursively shrink counter examples making them
smaller and smaller still.  Then I can go through all the code and make
it beautiful and clear (it's currently prototype hackiness).